requirement.txt
flask==2.3.2
dasbus==1.7
python-logging-loki==0.3.1